+++ include/core/Edge.h
